modal, terminacion de administracion de usarios y middleware

+++ resources/views/post/index.blade.php
@@ -7,19 +7,21 @@
 		<th>Nombre</th>
 		<th>Slug</th>
 		<th>Categoria</th>
-		<th>Opciones</th>
+		@if(Auth::user()->id_rol<2) <th>Opciones</th> @endif
 	</tr>
 	@foreach($posts as $post)
 		<tr>
 			<td>{{ $post->title }}</td>
 			<td>{{ $post->slug }}</td>
 			<td>{{ $post->name }}</td>
-			<td>
-				{!! Form::open(['route'=>['post.destroy',$post->id],'method'=>'DELETE']) !!}
-				{!! link_to_route('post.edit',$title='Editar',$parameters = $post->id) !!}
-				{!! Form::submit('Eliminar',['class'=>'btn btn-danger']) !!}
-				{!! Form::close() !!}
-			</td>
+			@if(Auth::user()->id_rol<2)
+				<td>
+					{!! Form::open(['route'=>['post.destroy',$post->id],'method'=>'DELETE']) !!}
+					{!! link_to_route('post.edit',$title='Editar',$parameters = $post->id) !!}
+					{!! Form::submit('Eliminar',['class'=>'btn btn-danger']) !!}
+					{!! Form::close() !!}
+				</td>
+			@endif
 		</tr>
 	@endforeach
 </table>
