+++ storage/framework/views/4c5c887adb8856ec884db518a726a0b0f2b224d9.php
@@ -1,22 +1,23 @@
 <?php $__env->startSection('contenido'); ?>
 	<table class="table table-hover">
-		<tr><th>Nombre</th><th>Opciones</th></tr>
+		<tr><th>Nombre</th><?php if(Auth::user()->id_rol<2): ?> <th>Opciones</th> <?php endif; ?></tr>
 		<?php $__currentLoopData = $cats; $__env->addLoop($__currentLoopData); foreach($__currentLoopData as $category): $__env->incrementLoopIndices(); $loop = $__env->getFirstLoop(); ?>
 			<tr>
 				<td><?php echo e($category->name); ?></td>
-				<td>
-				<?php echo link_to_route('categorias.edit',$title='Editar',$parameters=$category->id); ?>
+				<?php if(Auth::user()->id_rol<2): ?>
+					<td>
+					<?php echo link_to_route('categorias.edit',$title='Editar',$parameters=$category->id); ?>
 
-				 
-				 <?php echo Form::open(['route'=>['categorias.destroy',$category->id],'method'=>'DELETE']); ?>
+					 
+					 <?php echo Form::open(['route'=>['categorias.destroy',$category->id],'method'=>'DELETE']); ?>
 
-					<?php echo Form::submit('Eliminar',['class'=>'btn btn-danger']); ?>
+						<?php echo Form::submit('Eliminar',['class'=>'btn btn-danger']); ?>
 
-				<?php echo Form::close(); ?>
+					<?php echo Form::close(); ?>
 
-				
-			 	</td>
-
+					
+				 	</td>
+				<?php endif; ?>
 			</tr>
 		<?php endforeach; $__env->popLoop(); $loop = $__env->getFirstLoop(); ?>
 	</table>
